[DataGridPremium] Fix Excel export Web Worker demo not working in dev mode (#16517)
