--- docs/worker.ts
@@ -1,3 +1,0 @@
-import { setupExcelExportWebWorker } from '@mui/x-data-grid-premium';
-
-setupExcelExportWebWorker();
