+++ packages/x-data-grid-premium/src/setupExcelExportWebWorker.ts
@@ -1,0 +1,1 @@
+export { setupExcelExportWebWorker } from './hooks/features/export/serializer/setupExcelExportWebWorker';
